# Fix issue 21716: std.regex performance regression.