--- a/std/regex/internal/ir.d
+++ b/std/regex/internal/ir.d
@@ -544,37 +544,58 @@ class CtfeFactory(alias EngineType, Char, alias func) : GenericFactory!(EngineTy
     }
 }
 
+private auto defaultFactoryImpl(Char)(const ref Regex!Char re)
+{
+    import std.regex.internal.backtracking : BacktrackingMatcher;
+    import std.regex.internal.thompson : ThompsonMatcher;
+    import std.algorithm.searching : canFind;
+    static MatcherFactory!Char backtrackingFactory;
+    static MatcherFactory!Char thompsonFactory;
+    if (re.backrefed.canFind!"a != 0")
+    {
+        if (backtrackingFactory is null)
+            backtrackingFactory = new RuntimeFactory!(BacktrackingMatcher, Char);
+        return backtrackingFactory;
+    }
+    else
+    {
+        if (thompsonFactory is null)
+            thompsonFactory = new RuntimeFactory!(ThompsonMatcher, Char);
+        return thompsonFactory;
+    }
+}
+
+// Used to generate a pure wrapper for defaultFactoryImpl. Based on the example in
+// the std.traits.SetFunctionAttributes documentation.
+private auto assumePureFunction(T)(T t)
+{
+    if (isFunctionPointer!T)
+    {
+        enum attrs = functionAttributes!T | FunctionAttribute.pure_;
+        return cast(SetFunctionAttributes!(T, functionLinkage!T, attrs)) t;
+    }
+}
+
 // A workaround for R-T enum re = regex(...)
 template defaultFactory(Char)
 {
+    // defaultFactory is constructed as a safe, pure wrapper over defaultFactoryImpl.
+    // It can be faked as pure because the static mutable variables are used to cache
+    // the key and character matcher. The technique used avoids delegates and GC.
     @property MatcherFactory!Char defaultFactory(const ref Regex!Char re) @safe pure
     {
-        auto defaultFactoryImpl = () {
-            import std.regex.internal.backtracking : BacktrackingMatcher;
-            import std.regex.internal.thompson : ThompsonMatcher;
-            import std.algorithm.searching : canFind;
-            static MatcherFactory!Char backtrackingFactory;
-            static MatcherFactory!Char thompsonFactory;
-            if (re.backrefed.canFind!"a != 0")
-            {
-                if (backtrackingFactory is null)
-                    backtrackingFactory = new RuntimeFactory!(BacktrackingMatcher, Char);
-                return backtrackingFactory;
-            }
-            else
-            {
-                if (thompsonFactory is null)
-                    thompsonFactory = new RuntimeFactory!(ThompsonMatcher, Char);
-                return thompsonFactory;
-            }
-        };
+        static auto impl(const ref Regex!Char re)
+        {
+            return defaultFactoryImpl(re);
+        }
 
-        // this should be faked as pure because the static mutable variables are
-        // used to cache the created instance, like memoize
-        alias T = typeof(defaultFactoryImpl);
-        enum attrs = functionAttributes!T | FunctionAttribute.pure_;
-        return (() @trusted =>
-            (cast(SetFunctionAttributes!(T, functionLinkage!T, attrs)) defaultFactoryImpl))()();
+        static auto pureImpl(const ref Regex!Char re) @trusted
+        {
+            auto p = assumePureFunction(&impl);
+            return p(re);
+        }
+
+        return pureImpl(re);
     }
 }