This PR fixes a performance regression in std.regex introduced by PR #7678. Bug issue 
 is [2176](https://issues.dlang.org/show_bug.cgi?id=21716).

Note that the PR is for a merge to stable, rather than master, as this is a regression.

PR #7678 made changes to std.regex to provide API functions marked as `pure`. The APIs involved memoizing prior regex matchers for performance, so they are not `pure` from the compiler's perspective. The technique used to change the function attributes to `pure` had the effect of introducing delegates. This caused GC resulting in performance degradation. PR #7678 was released as part of DMD-2.095.

This PR modifies the code to use static functions rather than delegates, avoiding GC. This removed GC generation and brought performance back in line with DMD-2.094.2 in my tests.

The method of fixing the bug was identified by @acehreli and outlined in a [3/19 dlang forum post](https://forum.dlang.org/post/s32lu4$1qen$1@digitalmars.com). I wasn't able to use the exact method in the post, but it is similar.

Performance changes were validated one of my standard [tsv-filter](https://github.com/eBay/tsv-utils/tree/master/tsv-filter) regular expression benchmarks. This tests runs a regular expression match 14 million times (14 million lines, 1 test per line). See [tsv-utils 2018 comparative benchmarks](https://github.com/eBay/tsv-utils/blob/master/docs/comparative-benchmarks-2018.md#test-details) for details.

Run-time for the above test on a 16GB MacMini:
DMD-2.094.2:   9.54 seconds
DMD-2.095.1:  10.30 seconds
DMD-stable w/ PR:   9.57 seconds

The specific times of course vary run-to-run, but over many runs are quite stable. There is no material difference between the 2.094.1 and PR versions.

GC generation was validated using the same program using `"--DRT-gcopt=profile:1"`. No GC was observed in DMD-2.094.2. GC is incurred using DMD-2.095.1. GC is eliminated by this PR. Runs showing this are below.

```
$ ./bin.dmd-2.094.2/tsv-filter "--DRT-gcopt=profile:1" --count -H --regex 10:'[RD].*ION[0-2]' ~/Data/apps.fs.usda.gov_fia_datamart/TREE_GRM_ESTN_14mil.tsv
137993
	Number of collections:  0
	Total GC prep time:  0 milliseconds
	Total mark time:  0 milliseconds
	Total sweep time:  0 milliseconds
	Max Pause Time:  0 milliseconds
	Grand total GC time:  0 milliseconds
GC summary:    5 MB,    0 GC    0 ms, Pauses    0 ms <    0 ms

$ ./bin.dmd-2.095.1/tsv-filter "--DRT-gcopt=profile:1" --count -H --regex 10:'[RD].*ION[0-2]' ~/Data/apps.fs.usda.gov_fia_datamart/TREE_GRM_ESTN_14mil.tsv
137993
	Number of collections:  666
	Total GC prep time:  6 milliseconds
	Total mark time:  61 milliseconds
	Total sweep time:  6 milliseconds
	Max Pause Time:  0 milliseconds
	Grand total GC time:  74 milliseconds
GC summary:    5 MB,  666 GC   74 ms, Pauses   67 ms <    0 ms

$ ./bin.dmd-patch/tsv-filter "--DRT-gcopt=profile:1" --count -H --regex 10:'[RD].*ION[0-2]' ~/Data/apps.fs.usda.gov_fia_datamart/TREE_GRM_ESTN_14mil.tsv
137993
	Number of collections:  0
	Total GC prep time:  0 milliseconds
	Total mark time:  0 milliseconds
	Total sweep time:  0 milliseconds
	Max Pause Time:  0 milliseconds
	Grand total GC time:  0 milliseconds
GC summary:    5 MB,    0 GC    0 ms, Pauses    0 ms <    0 ms
```

I did not investigate the correctness of marking these APIs as `pure` introduced in PR #7678, and am assuming this was reviewed as part of that PR.