@jondegenhardt 2021-03-21
This is a regression, not an enhancement as marked by the dlang-bot. I don't seem to be able to change this.