@n8sh 2021-03-22
My problems with the implementation already existed before this patch. So I'm rescinding my requested changes and enabling `auto-merge` so a fix can get into stable before the next release.