@jondegenhardt 2021-03-22
> My problems with the implementation already existed before this patch. So I'm rescinding my requested changes and enabling `auto-merge` so a fix can get into stable before the next release.

Thank you. I appreciate your thoughtfulness about this. I too am not convinced the previous code was correct. At the same time, I haven't studied it enough to be definitive about where it is incorrect and exactly what the best solution going forward.